REFACTOR: Migrated blog/[id] page to typescript FEAT: Added scroll-to-top button to blog page
